+++ lib/rust/parser/generate-java/java/org/enso/syntax2/Parser.java
@@ -44,10 +44,11 @@ public final class Parser implements AutoCloseable {
       }
       System.load(path.getAbsolutePath());
     } catch (NullPointerException | IllegalArgumentException | LinkageError e) {
-      if (searchFromDirToTop(e, root, "target", "rust", "debug", name)) {
+      if (searchFromDirToTop(e, root, "target", "rust", "parser-jni", name)) {
         return;
       }
-      if (searchFromDirToTop(e, new File(".").getAbsoluteFile(), "target", "rust", "debug", name)) {
+      if (searchFromDirToTop(
+          e, new File(".").getAbsoluteFile(), "target", "rust", "parser-jni", name)) {
         return;
       }
       throw new IllegalStateException("Cannot load parser from " + root, e);